# Further tuning on flattenings of rings