--- a/test/Rings/MPolyAnyMap/flattenings.jl
+++ b/test/Rings/MPolyAnyMap/flattenings.jl
@@ -16,7 +16,7 @@
 
   R, _ = quo(R, ideal(R, sum(gens(R))))
   S, _ = polynomial_ring(R, [:s, :t])
-  phi = Oscar.flatten(S)
+  phi = Oscar.flatten(S; cached=true)
   @test vcat(phi.(gens(S)), Oscar.map_from_coefficient_ring_to_flattening(phi).(gens(coefficient_ring(S)))) == gens(codomain(phi))
   @test gens(S) == inverse(phi).(phi.(gens(S)))
   @test Oscar.map_from_coefficient_ring_to_flattening(phi).(gens(R)) == phi.(S.(gens(R)))
@@ -302,7 +302,7 @@ end
   R, (x,y,z) = polynomial_ring(QQ, [:x, :y, :z], cached=false)
   S, _ = polynomial_ring(R, [:s, :t], cached=false)
   S, _ = quo(S, ideal(S, elem_type(S)[]))
-  phi = Oscar.flatten(S)
+  phi = Oscar.flatten(S; cached=true)
   # julia might not consider `x^2+y^2` as being unused until the end of the scope
   # see https://github.com/JuliaLang/julia/issues/51818
   # so we put this into a function