Some more steps towards addressing #4112 . ~Still WIP but seemingly a speedup of factor 10 already. Timings will follow.~ I think, this is what I can do for the moment.