@HechtiDerLachs 2024-12-13
**Timings:** Again, I am using [this polynomial](https://github.com/user-attachments/files/18123872/poly.txt) as a benchmark.

before:
```julia
julia> f = load("/home/matthias/Downloads/poly.txt");

julia> g = copy(f);

julia> S = parent(f);

julia> Oscar.flatten(S);

julia> @time Oscar.flatten(S)(g);
  0.227188 seconds (1.05 M allocations: 46.635 MiB, 53.65% gc time)

julia> g = copy(f);

julia> @time Oscar.flatten(S)(g);
  0.099856 seconds (1.00 M allocations: 46.005 MiB)

julia> g = copy(f);

julia> @time Oscar.flatten(S)(g);
  0.112565 seconds (1.40 M allocations: 57.970 MiB)

julia> g = copy(f);

julia> @time Oscar.flatten(S)(g);
  0.244818 seconds (1.39 M allocations: 56.852 MiB, 46.52% gc time)
```
after:
```julia
julia> f = load("/path/to/poly.txt");

julia> g = copy(f);

julia> S = parent(f);

julia> Oscar.flatten(S);

julia> @time Oscar.flatten(S)(g);
  0.054546 seconds (578.67 k allocations: 27.150 MiB)

julia> g = copy(f);

julia> @time Oscar.flatten(S)(g);
  0.061309 seconds (578.67 k allocations: 27.150 MiB)

julia> g = copy(f);

julia> @time Oscar.flatten(S)(g);
  0.039068 seconds (428.25 k allocations: 20.028 MiB)
```
So roughly a factor 2 in improvement for the special cases covered compared to current master.