@thofma 2024-12-13
Are you caching the image of elements by default? Or why is the `copy` necessary in the benchmark?