@HechtiDerLachs 2024-12-13
On a first glance it seems to me that none of the failing tests is really my fault. Correct me if I'm wrong! 

Otherwise this should be good to go.