@HechtiDerLachs 2024-12-13
Picking up on the timings by @fingolfin on #4112 done for the tests of the `EllipticSurface`s:

current master: `310.391684 seconds (2.06 G allocations: 72.914 GiB, 12.26% gc time, 0.12% compilation time)`
with this PR: `310.526301 seconds (2.08 G allocations: 73.502 GiB, 12.21% gc time, 0.24% compilation time)` 

So either the expensive tasks lay elsewhere, or the improvements here just don't pay off as much as anticipated. Both are not unlikely, the first because flattening of rings only plays a role in the computation of gluings which is not the main objective of the algorithms tested there. For the second: We do not have more fine-tuned methods for polynomial- and quotient rings over localizations (yet). These come up naturally in the gluings as we usually glue over principal open subsets. I left a comment in the source code to indicate this as a potential thread for future work. But as this is not a super urgent matter at the moment, I would prefer to spend my resources differently for the moment.