@lgoettgens 2024-12-13
The `nightly, book` failure is due to https://github.com/oscar-system/Oscar.jl/issues/4402, so can be disregarded here.